Extract concrete resource implementations to dedicated module
Moved all concrete implementations from resource.rs to resource/implementations.rs
as part of breaking down the 1,415-line file. This extraction includes:

- Resource struct and core implementation
- ResourceScope for automatic cleanup
- TimedResourceGuard for timeout-based management
- ResourceLeakDetector and leak tracking
- ManagedResource wrapper
- ResourceExt trait and implementations

The resource.rs file now only contains error types, database module,
locking module, tests, and integration module. Further extractions
will continue to decompose these remaining components.

All tests pass (with single-threaded execution due to global state).
